+++ CHANGELOG.rst
@@ -12,6 +12,7 @@ Added
 
 Changed
 =======
+- The functionality `match_and_aply` has been added to match flows and apply actions. This matches a given packet against the stored flows.
 
 Deprecated
 ==========
@@ -21,6 +22,7 @@ Removed
 - Removed dependency from `amlight/scheduler`
 - Removed support for OpenFlow 1.0
 - Unsubscribe to the `amlight/flow_stats.flows_updated` event
+- Dependency on `flow_stats` has been removed, from where the functionality `match_and_aply` was previously used.
 
 Fixed
 =====