+++ CHANGELOG.rst
@@ -6,6 +6,19 @@ All notable changes to the sdntrace_cp NApp will be documented in this file.
 [UNRELEASED] - Under development
 ********************************
 
+Added
+=====
+
+Changed
+=======
+- ``PUT /traces`` will return the results in order, without aggregating them by `dpid`. Also, failed traces are not omitted.
+
+Removed
+=======
+
+Fixed
+=====
+
 [2022.3.0] - 2022-12-15
 ***********************