+++ CHANGELOG.rst
@@ -6,6 +6,9 @@ All notable changes to the sdntrace_cp NApp will be documented in this file.
 [UNRELEASED] - Under development
 ********************************
 
+[2022.3.1] - 2023-02-24
+***********************
+
 Added
 =====